Se agrega codigo a TurnoService,se crea clase TurnoController

+++ integradorClinica/src/main/java/com/backend/digitalhouse/integradorClinica/dto/salida/paciente/DomicilioSalidaDto.java
@@ -2,7 +2,7 @@ package com.backend.digitalhouse.integradorClinica.dto.salida.paciente;
 
 public class DomicilioSalidaDto {
 
-    private int id;
+    private long id;
     private String calle;
     private int numero;
     private String localidad;
@@ -11,7 +11,7 @@ public class DomicilioSalidaDto {
     public DomicilioSalidaDto() {
     }
 
-    public DomicilioSalidaDto(int id, String calle, int numero, String localidad, String provincia) {
+    public DomicilioSalidaDto(long id, String calle, int numero, String localidad, String provincia) {
         this.id = id;
         this.calle = calle;
         this.numero = numero;
@@ -19,7 +19,7 @@ public class DomicilioSalidaDto {
         this.provincia = provincia;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
